✨ feat(prisma): PrismaTypeWeaver enumType
