+++ packages/prisma/src/model-type-builder.ts
@@ -12,118 +12,8 @@ import {
   GraphQLScalarType,
   GraphQLString,
   GraphQLFloat,
-  type GraphQLFieldConfig,
 } from "graphql"
 import { capitalize } from "./utils"
-
-export class PrismaTypeWeaver {
-  protected modelMeta: Required<PrismaModelMeta>
-
-  constructor(modelData: PrismaModelMeta) {
-    this.modelMeta = PrismaTypeWeaver.indexModelMeta(modelData)
-  }
-
-  public inputType(name: string): GraphQLObjectType {
-    const input = this.modelMeta.inputTypes.get(name)
-    if (!input) throw new Error(`Input type ${name} not found`)
-    const existing = weaverContext.getNamedType(name) as
-      | GraphQLObjectType
-      | undefined
-
-    if (existing) return existing
-
-    return weaverContext.memoNamedType(
-      new GraphQLObjectType({
-        name,
-        fields: () => ({
-          ...Object.fromEntries(
-            input.fields.map((field) => {
-              const isNonNull = field.isRequired && !field.isNullable
-              const fieldInput = PrismaTypeWeaver.getMostRankInputType(
-                field.inputTypes
-              )
-
-              const isList = fieldInput.isList
-              let type: GraphQLOutputType = (() => {
-                switch (fieldInput.location) {
-                  case "inputObjectTypes":
-                    return this.inputType(fieldInput.type)
-                  case "scalar":
-                    return PrismaWeaver.getGraphQLTypeByField(fieldInput.type)!
-                  default:
-                    throw new Error(
-                      `Unknown input type location: ${fieldInput.location}`
-                    )
-                }
-              })()
-
-              if (isList) type = new GraphQLList(new GraphQLNonNull(type))
-              if (isNonNull) type = new GraphQLNonNull(type)
-
-              return [
-                field.name,
-                {
-                  description: field.comment,
-                  deprecationReason: field.deprecation?.reason,
-                  type,
-                } as GraphQLFieldConfig<any, any, any>,
-              ]
-            })
-          ),
-        }),
-      })
-    )
-  }
-
-  protected static getMostRankInputType(
-    inputTypes: readonly DMMF.InputTypeRef[]
-  ): DMMF.InputTypeRef {
-    const rankList = inputTypes
-      .map((item) => ({
-        item,
-        rank: PrismaTypeWeaver.getInputTypeRank(item),
-      }))
-      .sort((a, b) => b.rank - a.rank)
-
-    return rankList[0].item
-  }
-
-  protected static getInputTypeRank(inputType: DMMF.InputTypeRef): number {
-    let rank = inputType.type.length
-    if (inputType.isList) rank += 10
-    switch (inputType.location) {
-      case "scalar":
-        rank += 10
-        break
-      case "enumTypes":
-        rank += 20
-        break
-      case "fieldRefTypes":
-        rank -= 30
-        break
-      case "inputObjectTypes":
-        rank += 40
-    }
-    return rank
-  }
-
-  protected static indexModelMeta(
-    modelMeta: PrismaModelMeta
-  ): Required<PrismaModelMeta> {
-    modelMeta.inputTypes ??= PrismaTypeWeaver.indexInputTypes(modelMeta.schema)
-    return modelMeta as Required<PrismaModelMeta>
-  }
-
-  protected static indexInputTypes(
-    schema: DMMF.Schema
-  ): Map<string, DMMF.InputType> {
-    const map = new Map<string, DMMF.InputType>()
-    for (const inputType of schema.inputObjectTypes.prisma) {
-      map.set(inputType.name, inputType)
-    }
-    return map
-  }
-}
 
 export class PrismaModelTypeBuilder<
   TModalSilk extends PrismaModelSilk<any, any>,
