+++ packages/prisma/test/type-weaver.spec.ts
@@ -1,0 +1,59 @@
+import { describe, it, expect } from "vitest"
+import * as g from "./generated"
+import { PrismaTypeWeaver } from "../src"
+import { printType } from "graphql"
+
+describe("PrismaModelTypeWeaver", () => {
+  const typeWeaver = new PrismaTypeWeaver(g.User.data)
+  it("should be able to create a type weaver", () => {
+    expect(typeWeaver).toBeDefined()
+  })
+
+  describe("inputType", () => {
+    it("should be able to create WhereInput", () => {
+      const UserWhereInput = typeWeaver.inputType("UserWhereInput")
+      expect(printType(UserWhereInput)).toMatchInlineSnapshot(`
+        "type UserWhereInput {
+          AND: [UserWhereInput!]
+          OR: [UserWhereInput!]
+          NOT: [UserWhereInput!]
+          id: IntFilter
+          email: StringFilter
+          name: StringNullableFilter
+          posts: PostListRelationFilter
+          publishedPosts: PostListRelationFilter
+          Profile: ProfileNullableRelationFilter
+        }"
+      `)
+    })
+
+    it("should be able to create ScaleFilter", () => {
+      const IntFilter = typeWeaver.inputType("IntFilter")
+      expect(printType(IntFilter)).toMatchInlineSnapshot(`
+      "type IntFilter {
+        equals: Int
+        in: [Int!]
+        notIn: [Int!]
+        lt: Int
+        lte: Int
+        gt: Int
+        gte: Int
+        not: NestedIntFilter
+      }"
+    `)
+    })
+  })
+
+  describe("enumType", () => {
+    it("should be able to create enum type", () => {
+      const SortOrder = typeWeaver.enumType("SortOrder")
+
+      expect(printType(SortOrder)).toMatchInlineSnapshot(`
+        "enum SortOrder {
+          asc
+          desc
+        }"
+      `)
+    })
+  })
+})
